+++ migrations/1690277351447222_real_estates.postgres.up.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS real_estates (
     item_id INTEGER REFERENCES items(id) ON DELETE CASCADE,
     title TEXT NOT NULL,
     type_id TEXT NOT NULL,
-    square_area INTEGER NOT NULL,
+    square_area FLOAT NOT NULL,
     land_serial_number TEXT NOT NULL,
     estate_serial_number TEXT NOT NULL,
     ownership_type TEXT NOT NULL,
